+++ CHANGES.rst
@@ -15,6 +15,7 @@
 - Menu: changed label of toolbox item to "3Di Results Analysis"
 - Result browser tab: add buttons are enabled again after addition of grid or result (#881)
 - Animation: Legend subgroups (#769)
+- Watershed: adding catchment result when animation styling is selected results no longer results in error (#883)
 
 
 3.1.6 (2023-07-20)